Fix the recovery scenarios on a host without a toc account, and a cast
The scenarios I added went red on CI while passing locally, for a reason
worth recording: the recovery script installs the rebuilt binary with
`install -o toc -g toc`, and a GitHub runner has no toc account, so the
install failed and the live binary was never replaced. It passed on my
machine only because that WSL instance had a leftover toc user from
earlier test servers -- an environment-specific false pass, which is
exactly what these scenarios exist to prevent.

Ownership is now TOC_OWNER/TOC_GROUP, defaulting to toc so production
behaviour is unchanged, and the scenarios install as whoever runs them.

Also silences a -Wcast-qual warning the strict build reported in the MCCP
path. zlib types next_in as a plain Bytef * even though deflate() only
reads it, so assigning a const input discards the qualifier; routing
through uintptr_t keeps the strict build clean rather than carrying a
warning for an API wart.

+++ src/telnet_proto.c
@@ -132,7 +132,13 @@ bool telnet_write_compressed( DESCRIPTOR_DATA *d, const char *txt, int length )
     if ( remaining == 0 )
         return true;
 
-    stream->next_in  = (Bytef *) txt;
+    /*
+     * zlib types next_in as a plain Bytef * even though deflate() only reads
+     * it, so a const input cannot be assigned without discarding the
+     * qualifier. Route through uintptr_t so the strict-warning build stays
+     * clean rather than carrying a -Wcast-qual warning for an API wart.
+     */
+    stream->next_in  = (Bytef *)(uintptr_t) txt;
     stream->avail_in = (uInt) remaining;
 
     while ( stream->avail_in > 0 )
